feat : 정산생성 2단계에서의 채팅방 view 제작
